I'm renaming movie-router-2.js to movie-router.js since is the only router in the app at the moment (the previous one was removed)
